pyproject.toml
[project]
name = "loongrid"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.11"
dependencies = [
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.6",
    "fastapi-cli>=0.0.6",
    "pydantic>=2.10.3",
    "sqlalchemy[asyncio]>=2.0.36",
    "pydantic-settings>=2.6.1",
    "psutil>=6.1.0",
    "cachetools>=5.5.0",
    "passlib[argon2]>=1.7.4",
    "email-validator>=2.2.0",
    "pyjwt>=2.10.1",
    "cryptography>=44.0.0",
    "pydantic-extra-types>=2.10.1",
    "phonenumbers>=8.13.52",
    "pyspellchecker>=0.8.2",
    "pycountry>=24.6.1",
    "boto3>=1.36.16",
]
